Ensure recursive field annotation resolution
At the end of the registry initialisation, we need to walk across all the fields defined on extendable models to replace the annoted orignal type by the one build into the registry if the field reference a extendable model. Once we replace the declared type buy the resolved one, we also need to rebuild the model schema to take into account this change. Prior to this change, the result was consistent. Indeed, the resolution mechanism was not applied recursively. As result the model rebuild for a class with a field declared as an extendable model type, could not contain  the resolved definition of fields declared into this referenced extendable model type if the resolution of the last one was not already done. The resolution mechanism is now recursive and when a annotation is resolved, we ensure that the new type is also resolved.